:ambulance:

+++ app/src/main/java/yoloyoj/pub/models/Event.kt
@@ -8,8 +8,8 @@ class Event(
     var like: Like? = null,
     var authorid: Int? = null,
     var place: String? = null,
-    var lat: Int? = null,
-    var lng: Int? = null
+    var lat: Double? = null,
+    var lng: Double? = null
 ) {
     override fun toString(): String {
         return "{ \n" +
